Change rules good accounts under SSO
Now that we're using SSO, the rules for "safe account" are more complex.
They are described in the _epilog string near the top of the file, which
is also output when the --help option is passed.

There are improvements to be made, but this is good enough for a first
pass.